requirements.txt
fastapi[standard]
sqlalchemy
pyjwt